+++ examples/assets/js/uix-kit.js
@@ -7,7 +7,7 @@
  * ## Project Name        :  Uix Kit Demo
  * ## Project Description :  Free Responsive HTML5 UI Kit for Fast Web Design Based On Bootstrap
  * ## Based on            :  Uix Kit (https://github.com/xizon/uix-kit)
- * ## Version             :  1.1.96
+ * ## Version             :  1.1.97
  * ## Last Update         :  April 2, 2018
  * ## Powered by          :  UIUX Lab
  * ## Created by          :  UIUX Lab (https://uiux.cc)
@@ -44,28 +44,29 @@
     19. Custom Core Scripts & Stylesheets 
     20. Bulleted List 
     21. Posts List With Ajax 
-    22. Fullwidth List of Split 
-    23. Mobile Menu 
-    24. Modal Dialog 
-    25. Multiple Items Carousel 
+    22. Fullwidth List of Split 2 
+    23. Fullwidth List of Split 
+    24. Mobile Menu 
+    25. Modal Dialog 
     26. Mousewheel Interaction 
-    27. Navigation Highlighting 
-    28. Parallax 
-    29. Periodical Scroll 
-    30. Pricing 
-    31. Progress Bar 
-    32. Retina Graphics for Website 
-    33. Scroll Reveal 
-    34. Show More Less 
-    35. Custom Lightbox 
-    36. Slideshow ( with custom flexslider ) 
-    37. Source Code 
-    38. Sticky Elements 
-    39. Tabs 
-    40. Testimonials Carousel 
-    41. Text effect 
-    42. Timeline 
-    43. AJAX 
+    27. Multiple Items Carousel 
+    28. Navigation Highlighting 
+    29. Parallax 
+    30. Periodical Scroll 
+    31. Pricing 
+    32. Progress Bar 
+    33. Retina Graphics for Website 
+    34. Scroll Reveal 
+    35. Show More Less 
+    36. Custom Lightbox 
+    37. Slideshow ( with custom flexslider ) 
+    38. Source Code 
+    39. Sticky Elements 
+    40. Tabs 
+    41. Testimonials Carousel 
+    42. Text effect 
+    43. Timeline 
+    44. AJAX 
 
 
 */
@@ -5105,6 +5106,73 @@ theme = ( function ( theme, $, window, document ) {
 
 /* 
  *************************************
+ * <!-- Fullwidth List of Split 2 -->
+ *************************************
+ */
+theme = ( function ( theme, $, window, document ) {
+    'use strict';
+   
+   
+    var pageLoaded = function() {
+		
+		var $window      = $( window ),
+			windowWidth  = $window.width(),
+			windowHeight = $window.height();
+		
+		
+		fullwidthListSplitInit( windowWidth );
+		
+		$window.on( 'resize', function() {
+			// Check window width has actually changed and it's not just iOS triggering a resize event on scroll
+			if ( $window.width() != windowWidth ) {
+
+				// Update the window width for next time
+				windowWidth = $window.width();
+
+				// Do stuff here
+				fullwidthListSplitInit( windowWidth );
+		
+
+			}
+		});
+		
+		
+		
+		function fullwidthListSplitInit( w ) {
+			
+			$( '.list-split-content-container' ).each(function() {
+		
+				if ( imgH > 0 ) {
+					$( this ).find( '.to-left, .to-right' ).css( 'height', imgH + 'px' );
+				}
+
+				if ( w <= 768 ) {
+					$( this ).find( '.to-left, .to-right' ).css( 'height', 'auto' );
+				}
+
+			});		
+			
+		}
+		
+
+		
+		
+	};
+		
+    theme.fullwidthListSplit2 = {
+        pageLoaded : pageLoaded        
+    };
+
+    theme.components.pageLoaded.push( pageLoaded );
+    return theme;
+
+}( theme, jQuery, window, document ) );
+
+
+
+
+/* 
+ *************************************
  * <!-- Fullwidth List of Split -->
  *************************************
  */
@@ -7408,10 +7476,34 @@ theme = ( function ( theme, $, window, document ) {
 			} 
 			
 			
+			// Determine if this slider is added with a synchronization event
+			$( '[data-mysync]' ).each( function()  {
+				var curSync      = $( this ).data( 'mysync' ),
+					thisSliderID = $this.attr( 'id' );
+				
+				
+				if( typeof curSync != typeof undefined ) {
+					curSync = curSync.toString().replace( '#', '' ).replace( '.', '' );
+				}
+				
+				if( typeof thisSliderID != typeof undefined && thisSliderID == curSync ) {
+					dataAuto = false;
+					dataPaging = false;
+					
+					// break out of jQuery each Loop
+					return false; 
+				}
+
+
+			});
+			
+			
 			//Display counter
 			var $countTotal = ( dataCountTotal ) ? $( dataCountTotal ) : $( 'p.count em.count' ), 
 				$countCur   = ( dataCountCur ) ? $( dataCountCur ) : $( 'p.count em.current' );
 			
+			
+
 			
 			$this.flexslider({
 				namespace	      : 'custom-theme-flex-',