+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-#### = 1.1.96 (April 2, 2018) =
+#### = 1.1.97 (April 2, 2018) =
 
 * New: Extend the module "Slideshow" that add a new demo "Sync Multiple Sliders".
-
+* New: Add a new module "Fullwidth List of Split 2".
 
 
 #### = 1.1.95 (March 30, 2018) =
@@ -166,7 +166,7 @@
 
 #### = 1.1.4 (February 23, 2018) =
 
-* New: Add a new module "List of Split With Full-Width".
+* New: Add a new module "Fullwidth List of Split".
 * Tweak: Modify ".container" width as a dynamic percentage.
 * Tweak: Optimized CSS code structure.